+++ relay-server/src/main/java/com/airh/relay/RelayServerApplication.java
@@ -2,7 +2,9 @@ package com.airh.relay;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
 @SpringBootApplication
 public class RelayServerApplication {
     public static void main(String[] args) {